Core/Misc: Fixed wrong db field accessors usage found with TRINITY_STRICT_DATABASE_TYPE_CHECKS
* Fixes loading group instance reset times
* Fixes loading battlefield templates

+++ src/server/game/Battlefield/BattlefieldMgr.cpp
@@ -53,7 +53,7 @@ void BattlefieldMgr::InitBattlefield()
         {
             Field* fields = result->Fetch();
 
-            uint32 typeId = fields[0].GetUInt32();
+            uint32 typeId = fields[0].GetUInt8();
 
             if (typeId >= BATTLEFIELD_MAX)
             {
